Develop



查看合并请求 !189

+++ workflow-admin/src/main/java/com/hengtian/common/utils/PageInfo.java
@@ -13,7 +13,7 @@ public class PageInfo {
 
     private final static int PAGESIZE = 10; //默认显示的记录数 
 
-    private int total; // 总记录 
+    private long total; // 总记录
     private List rows; //显示的记录  
 
     @JsonIgnore
@@ -77,11 +77,11 @@ public class PageInfo {
         this.order = order;
     }
 
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(long total) {
         this.total = total;
     }
